fix: update error handling in KeyStore.
